+++ decorator/src/com/david/arutiunian/lab2/input/InputStreamImpl.kt
@@ -1,7 +1,7 @@
 package com.david.arutiunian.lab2.input
 
 abstract class InputStreamImpl : InputStream {
-    override fun readBlock(buffer: ByteArray, size: StreamSize): StreamSize {
+    override fun readBlock(buffer: IntArray, size: StreamSize): StreamSize {
         var length = size
         var index = 0
