+++ decorator/src/com/david/arutiunian/lab2/input/MemoryInputStream.kt
@@ -2,12 +2,12 @@ package com.david.arutiunian.lab2.input
 
 import java.io.IOException
 
-class MemoryInputStream(private val stream: ByteArray) : InputStreamImpl() {
+class MemoryInputStream(private val stream: IntArray) : InputStreamImpl() {
     private var position: StreamSize = 0
 
     override fun isEOF() = position == stream.size
 
-    override fun readByte(): Byte {
+    override fun readByte(): Int {
         if (position >= stream.size) {
             throw IOException("Failed to load next byte at $position")
         }
