+++ decorator/src/com/david/arutiunian/lab2/output/MemoryOutputStream.kt
@@ -2,10 +2,10 @@ package com.david.arutiunian.lab2.output
 
 import java.io.IOException
 
-class MemoryOutputStream(private val stream: ByteArray) : OutputStreamImpl() {
+class MemoryOutputStream(private val stream: IntArray) : OutputStreamImpl() {
     private var position: StreamSize = 0
 
-    override fun writeByte(byte: Byte) {
+    override fun writeByte(byte: Int) {
         if (position >= stream.size) {
             throw IOException("Failed to write next byte at $position")
         }
