+++ decorator/src/com/david/arutiunian/lab2/output/OutputStream.kt
@@ -3,7 +3,7 @@ package com.david.arutiunian.lab2.output
 typealias StreamSize = Int
 
 interface OutputStream {
-    fun writeByte(byte: Byte)
+    fun writeByte(byte: Int)
 
-    fun writeBlock(source: ByteArray, size: StreamSize)
+    fun writeBlock(source: IntArray, size: StreamSize)
 }
