+++ decorator/src/com/david/arutiunian/lab2/output/OutputStreamImpl.kt
@@ -1,7 +1,7 @@
 package com.david.arutiunian.lab2.output
 
 abstract class OutputStreamImpl : OutputStream {
-    override fun writeBlock(source: ByteArray, size: StreamSize) {
+    override fun writeBlock(source: IntArray, size: StreamSize) {
         var length = size
         var index = 0
